Manage new Firefox profiles.ini file format. Fix #25.

+++ src/main/kotlin/bandcampcollectiondownloader/CookiesManagement.kt
@@ -40,8 +40,8 @@ object CookiesManagement {
     )
 
     data class Cookies (
-        val source: Path,
-        val content: Map<String,String>
+            val source: Path,
+            val content: Map<String,String>
     )
 
 
@@ -68,20 +68,22 @@ object CookiesManagement {
         val parsedCookies =
                 try {
                     if (cookiesFile.toString().endsWith(".json"))
-                          parsedCookiesToMap(gson.fromJson(fileData, Array<ParsedCookie>::class.java))  else
-                          parseCookiesText(fileData)
+                        parsedCookiesToMap(gson.fromJson(fileData, Array<ParsedCookie>::class.java))  else
+                        parseCookiesText(fileData)
                 } catch (e: JsonSyntaxException) {
                     throw BandCampDownloaderError("Cookies file '$cookiesFile' is not well formed: ${e.message}")
                 }
         return Cookies(cookiesFile, parsedCookies)
     }
 
-    fun retrieveFirefoxCookies(): Cookies {
-        val result = HashMap<String, String>()
+    /**
+     * Searches in all Firefox profiles for valid Bandcamp cookies, and returns all found sets of cookies.
+     */
+    fun retrieveFirefoxCookies(): List<Cookies> {
 
-        // Find cookies file path
+        val allFoundCookies = ArrayList<Cookies>()
 
-
+        // Find Firefox configuration folder
         val firefoxConfDirPath: Path?
         firefoxConfDirPath = when {
             Util.isUnix() -> {
@@ -95,51 +97,65 @@ object CookiesManagement {
             else -> throw BandCampDownloaderError("OS not supported, cannot find Firefox cookies!")
         }
 
+        // Find all firefox profiles
         val profilesListPath = firefoxConfDirPath.resolve("profiles.ini")
         val profilesListFile = File(profilesListPath.toUri())
         if (!profilesListFile.exists()) {
             throw BandCampDownloaderError("No Firefox profiles.ini file could be found!")
         }
         val ini = Ini(profilesListFile)
-        val default = "Default"
-        val defaultProfileSection = ini.keys.find {
-            ini[it] != null
-                    && ini[it]!!.containsKey(default)
-                    && ini[it]!![default] == "1"
+        val pathKey = "Path"
+        val entriesWithPath = ini.keys.filter {
+            ini[it] != null && ini[it]!!.containsKey(pathKey)
         }
-        val defaultProfilePath = firefoxConfDirPath.resolve(ini.get(defaultProfileSection, "Path"))
-        val cookiesFilePath = defaultProfilePath.resolve("cookies.sqlite")
 
-        // Copy cookies file as tmp file
-        val tmpFolder = Files.createTempDirectory("bandcampCollectionDownloader")
-        val copiedCookiesPath = Files.copy(cookiesFilePath, tmpFolder.resolve("cookies.json"))
-        copiedCookiesPath.toFile().deleteOnExit()
+        // For each profile, look for cookies
+        for (entryWithPath in entriesWithPath) {
 
-        // Start reading firefox's  cookies.sqlite
-        var connection: Connection? = null
-        try {
-            // create a database connection
-            connection = DriverManager.getConnection("jdbc:sqlite:$copiedCookiesPath")
-            val statement = connection!!.createStatement()
-            statement.queryTimeout = 30  // set timeout to 30 sec.
-            val rs = statement.executeQuery("select * from moz_cookies where host = '.bandcamp.com'")
-            // For each resulting row
-            while (rs.next()) {
-                // Extract data from row
-                val name = rs.getString("name")
-                val value = rs.getString("value")
-                val expiry = rs.getString("expiry").toLong()
+            val result = HashMap<String, String>()
 
-                // We only keep cookies that have not expired yet
-                val now = Instant.now().epochSecond
-                val difference = expiry - now
-                if (difference > 0)
-                    result[name] = value
+            val profilePath = firefoxConfDirPath.resolve(ini.get(entryWithPath, pathKey))
+            val cookiesFilePath = profilePath.resolve("cookies.sqlite")
+
+            if (Files.exists(cookiesFilePath)) {
+
+                // Copy cookies file as tmp file
+                val tmpFolder = Files.createTempDirectory("bandcampCollectionDownloader")
+                val copiedCookiesPath = Files.copy(cookiesFilePath, tmpFolder.resolve("cookies.json"))
+                copiedCookiesPath.toFile().deleteOnExit()
+
+                // Start reading firefox's  cookies.sqlite
+                var connection: Connection? = null
+                try {
+                    // create a database connection
+                    connection = DriverManager.getConnection("jdbc:sqlite:$copiedCookiesPath")
+                    val statement = connection!!.createStatement()
+                    statement.queryTimeout = 30  // set timeout to 30 sec.
+                    val rs = statement.executeQuery("select * from moz_cookies where host = '.bandcamp.com'")
+                    // For each resulting row
+                    while (rs.next()) {
+                        // Extract data from row
+                        val name = rs.getString("name")
+                        val value = rs.getString("value")
+                        val expiry = rs.getString("expiry").toLong()
+
+                        // We only keep cookies that have not expired yet
+                        val now = Instant.now().epochSecond
+                        val difference = expiry - now
+                        if (difference > 0)
+                            result[name] = value
+                    }
+                } finally {
+                    connection?.close()
+                }
+                if (result.isNotEmpty()) {
+                    allFoundCookies.add(Cookies(cookiesFilePath, result))
+                }
             }
-        } finally {
-            connection?.close()
+
         }
-        return Cookies(cookiesFilePath, result)
+        return allFoundCookies
     }
+
 
 }
